+++ src/rgw/driver/sfs/CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fixed segfault when SFSAtomicWriter::complete is called whith mtime output
+  variable set to nullptr
+
 ## [0.7.0] - 2022-10-20
 
 ### Fixed